DEPLOY/backend/requirements.txt
# Voice AI Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# LLM Client (OpenAI-compatible for Ollama)
openai>=1.0.0

# Text-to-Speech
edge-tts>=6.1.0

# Environment Management
python-dotenv>=1.0.0

# Desktop Application
pywebview>=5.0.0

# Async Support
aiohttp>=3.9.0
anyio>=4.0.0

# Optional: System Tray (Windows)
# pystray>=0.19.0
# pillow>=10.0.0

# Development Dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
